% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protgenerics.R
\name{extractByIndex}
\alias{extractByIndex}
\title{Extracting elements by index}
\usage{
extractByIndex(object, i)
}
\arguments{
\item{object}{The object to subset/from which to extract elements.}

\item{i}{\code{integer} with the indices.}
}
\description{
The \code{extractByIndex()} method allows to subset an object (or extract
elements from it) by providing their \code{integer} indices.
}
