% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protgenerics.R
\name{filterFeatures}
\alias{filterFeatures}
\title{Filter features}
\usage{
filterFeatures(object, filter, ...)
}
\arguments{
\item{object}{The object to filter.}

\item{filter}{The filtering criteria on which \code{object} should be filtered.}

\item{...}{Optional parameters.}
}
\description{
Implementations of this generic filter function are supposed to filter
\emph{features} in \code{object} based on a filter criteria defined by
parameter \code{filter}.
}
