% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protgenerics.R
\name{filterSpectra}
\alias{filterSpectra}
\title{Filter Spectra}
\usage{
filterSpectra(object, filter, ...)
}
\arguments{
\item{object}{The object to filter.}

\item{filter}{The filtering criteria based on which \code{object} should be
filtered.}

\item{...}{Optional parameters.}
}
\description{
Implementations of this generic filter function are supposed to filter
\emph{spectra} (e.g. within a \code{Spectra} object) based on filter criteria
defined with parameter \code{filter}.
}
