% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataClasses.R
\docType{class}
\name{Param}
\alias{Param}
\alias{Param-class}
\alias{as.list,Param-method}
\alias{show,Param-method}
\title{Generic parameter class}
\usage{
\S4method{as.list}{Param}(x, ...)

\S4method{show}{Param}(object)
}
\arguments{
\item{x}{`Param` object.}

\item{...}{ignored.}

\item{object}{`Param` object.}
}
\description{
The `Param` class is a virtual class which can be used as *base* class
from which *parameter* classes can inherit.

The methods implemented for the `Param` class are:

- `as.list`: coerces the `Param` class to a `list` with list elements
  representing the object's slot values, names the slot names. *Hidden*
  slots (i.e. those with a name starting with `.`) are not returned. In
  addition, a `Param` class can be coerced to a `list` using
  `as(object, "list")`.

- `show`: prints the content of the `Param` object (i.e. the individual
  slots and their value).
}
\author{
Johannes Rainer
}
