% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protgenerics.R
\name{peaksData}
\alias{peaksData}
\alias{peaksData<-}
\alias{peaksVariables}
\title{Get or set MS peak data}
\usage{
peaksData(object, ...)

peaksData(object) <- value

peaksVariables(object, ...)
}
\arguments{
\item{object}{A data object.}

\item{...}{Optional parameters.}

\item{value}{Replacement for peaks data.}
}
\description{
These methods get or set mass spectrometry (MS) peaks data, which can be
m/z, intensity or retention time values.
See the respective help pages (e.g. in the \code{Spectra} or \code{Chromatograms}
packages) for information on the actual implementations of these methods.
}
